Restricted amount of moves per turn, added a way to win and a game over scene
Each player gets 5 moves per turn and can build them up if they choose. The first player to reach 5 points (gold) wins and it goes to the game over screen. The screen displays each player's score and has the option to play again or go back to the main menu.